Moved BBOP graph stuff to separate file
