+++ src/api/BioLink.js
@@ -3,6 +3,7 @@ import us from 'underscore';
 import * as bbopgraph from 'bbop-graph';
 import { labelToId, isTaxonCardType } from '../lib/TaxonMap';
 import getStaticSourceData from './StaticSourceData';
+import * as bbopgraphUtil from './BBOPGraphUtil';
 
 // Example of a domain-specific (as opposed to a generic loadJSON)
 // service function. This set of domain-specific services will pretty much
@@ -357,81 +358,17 @@ export async function getSources() {
     .value();
 
   // put things from dynamic data into sourceData
-  _populateIngestDate(sourceData, dynamicSourceDataGraph);
-  _populateRdfDownloadUrl(sourceData, dynamicSourceDataGraph);
-  _populateSourceFiles(sourceData, dynamicSourceDataGraph);
+  bbopgraphUtil._populateIngestDate(sourceData, dynamicSourceDataGraph);
+  bbopgraphUtil._populateRdfDownloadUrl(sourceData, dynamicSourceDataGraph);
+  bbopgraphUtil._populateSourceFiles(sourceData, dynamicSourceDataGraph);
 
   // We still need static data for some things, e.g. source display name, text descriptions of each source,
   // and usage, since these aren't in Scigraph (and possibly shouldn't be). Any item in staticSourceData will overwrite
   // item from dynamic data, to allow us to override stuff from db using static data
   const staticSourceData = getStaticSourceData();
-  sourceData = _mergedStaticData(sourceData, staticSourceData);
+  sourceData = bbopgraphUtil._mergedStaticData(sourceData, staticSourceData);
 
   return sourceData;
-}
-
-function _mergedStaticData(sourceData, staticSourceData) {
-  sourceData = us.map(sourceData, function (sourceDatum) {
-    if (staticSourceData.hasOwnProperty(sourceDatum._summary_iri)) {
-      const staticDatum = staticSourceData[sourceDatum._summary_iri];
-      return Object.assign({}, sourceDatum, staticDatum);
-    }
-    return sourceDatum;
-
-  });
-  return sourceData;
-}
-
-function _populateSourceFiles(sourceData, graph) {
-  for (let i = 0; i < sourceData.length; i++) {
-    const sources = _subjectPredicate2Objects(sourceData[i]._version_iri, 'dcterms:source', graph);
-    sourceData[i].sourceFiles = us.chain(sources)
-      .map(function (source) {
-        const node = graph.get_node(source);
-        let retVal = 'Unknown';
-        if (node._metadata.hasOwnProperty('http://purl.org/pav/retrievedOn')) {
-          retVal = node._metadata['http://purl.org/pav/retrievedOn'][0];
-        }
-        return { 'fileUrl': source, 'retrievedOn': retVal };
-      })
-      .value();
-  }
-}
-
-function _populateIngestDate(sourceData, graph) {
-  for (let i = 0; i < sourceData.length; i++) {
-    sourceData[i].ingestDate =
-          graph.get_node(sourceData[i]._version_iri)._metadata['http://purl.org/dc/terms/created'][0];
-  }
-}
-
-function _populateRdfDownloadUrl(sourceData, graph) {
-  for (let i = 0; i < sourceData.length; i++) {
-    const distribution_iri = _versionIRI2distributionIRI(sourceData[i]._version_iri, graph);
-    const downloadUrls = _subjectPredicate2Object(distribution_iri, 'dcterms:downloadURL', graph);
-    sourceData[i].rdfDownloadUrl = downloadUrls.replace('MonarchArchive:', 'https://archive.monarchinitiative.org/');
-  }
-}
-
-function _versionIRI2distributionIRI(version_iri, graph) {
-  return _subjectPredicate2Object(version_iri, 'dcat:Distribution', graph);
-}
-
-function _subjectPredicate2Objects(subject_iri, predicate, graph) {
-  const edges = graph.get_edges_by_subject(subject_iri);
-  const object_iri = us.chain(edges)
-    .filter(function (edge) {
-      return edge._predicate_id == predicate;
-    })
-    .pluck('_object_id')
-    .value();
-  return object_iri;
-}
-
-function _subjectPredicate2Object(subject_iri, predicate, graph) {
-  return us.chain(_subjectPredicate2Objects(subject_iri, predicate, graph))
-    .first()
-    .value();
 }
 
 export async function getSearchResults(query, start, rows, categories, taxa) {
